requirements.txt
torch
torchvision
torchaudio
numpy
pandas
matplotlib
pillow
scikit-image
kagglehub
argparse
tqdm
